Fixed encryption of sensitive data
Changed encryption method to OpenSSL SHA256 hash algorithm as recommended by my Professor. I downloaded the library and included the dependencies in the CMakeList file for the project. I am using the SHA256 method to encrypt a string by reference for storage in a database. Project is running successfully and encryption is working successfully.