Added Input Validation / Sanitization
Added input validation and sanitization to the process of gathering input from the user for creating a new account. For each field, the input is validated for length requirements and sanitized to removed unexpected characters for certain fields (i.e. no numbers allowed in First Name). Next will clean up DatabaseConnection methods and implement prepared statements to create a new account in the DB for this process with logging for DB changes.